# lab4/docs/swagger.yaml
basePath: /api/v1
definitions:
  serializer.CurrentDeviceJSON:
    properties:
      amount:
        type: integer
      amperage:
        type: number
      curr_dev_id:
        type: integer
      current_id:
        type: integer
      device_id:
        type: integer
    type: object
  serializer.CurrentJSON:
    properties:
      created_at:
        type: string
      creator_login:
        type: string
      current_id:
        type: integer
      finish_date:
        type: string
      form_date:
        type: string
      moderator_login:
        type: string
      status:
        type: string
      voltage_bord:
        type: number
    type: object
...
host: localhost
info:
  contact:
    email: support@current.com
    name: API Support
    url: http://localhost
  description: API для определения необходимой силы тока
  license:
    name: MIT
  title: Current API
  version: "1.0"
paths:
...
  /current-devices/{current_id}/{device_id}:
    delete:
      description: Удаляет связь устройства и заявки
      parameters:
      - description: ID устройства
        in: path
        name: device_id
        required: true
        type: integer
      - description: ID заявки
        in: path
        name: current_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Обновленная заявка
          schema:
            $ref: '#/definitions/serializer.CurrentJSON'
        "400":
          description: Неверные ID
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещен
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Не найдено
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Удалить устройство из заявки
      tags:
      - current_devices
    put:
      consumes:
      - application/json
      description: Обновляет параметры устройства в конкретной заявке
      parameters:
      - description: ID устройства
        in: path
        name: device_id
        required: true
        type: integer
      - description: ID заявки
        in: path
        name: current_id
        required: true
        type: integer
      - description: Новые данные
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/serializer.CurrentDeviceJSON'
      produces:
      - application/json
      responses:
        "200":
          description: Обновленные данные
          schema:
            $ref: '#/definitions/serializer.CurrentDeviceJSON'
        "400":
          description: Неверные данные
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Не найдено
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Изменить данные устройства в заявке
      tags:
      - current_devices
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"